fix: move read_gradle_json and clone_p4a to module
