+++ kvdeveloper/module.py
@@ -4,6 +4,11 @@ import platform
 import subprocess
 import sys
 import re
+
+import io
+import json
+import requests
+
 from shutil import rmtree
 from typing import Dict, List
 from kvdeveloper.utils import (
@@ -981,3 +986,55 @@ def remove_from_structure(
                     console.print(f"\nDeleted: [red]{controller_file_path}[/red]")
             except Exception as e:
                 console.print(f"Error: [red]{e}[/red]")
+
+def read_gradle_json(path: str):
+    """
+    gradle.json from path to dict
+    :param path: Path of the gradle.json
+    :return:
+        Python dict with the keywords ['classpath', 'plugin', 'bom', 'dep']
+        For each of the keys it does set(value) to avoid duplicates.
+    """
+    gradle_json = {
+        'classpath': [],
+        'plugin': [],
+        'bom': [],
+        'dep': []
+        }
+
+    if os.path.exists(path):
+        with open(path, "r", encoding="UTF-8") as file:
+            g_json = json.load(file)
+            for k, v in g_json.items():
+                if gradle_json.get(k) is not None:
+                    gradle_json[k].extend(v)
+                    gradle_json[k] = list(set(gradle_json[k]))
+
+                else:
+                    raise ValueError(
+                            "Key does not belong to {gradle_json.keys()}"
+                            )
+    return gradle_json
+
+def clone_p4a(p4a_dir: str, url: str):
+    with requests.get(url, stream=True) as response:
+        response.raise_for_status()
+        with open(f"{p4a_dir}.tar.gz", 'wb') as file:
+            for chunk in response.iter_content(
+                    chunk_size=io.DEFAULT_BUFFER_SIZE): 
+                file.write(chunk)
+
+    console.print(f"{p4a_dir} Creating it...")
+    os.mkdir(p4a_dir)
+    untar_command = [
+            "tar", "-xvzf", "python-for-android.tar.gz",
+            "-C", "python-for-android",
+            "--strip-components=1"
+            ]
+
+    console.print(' '.join(untar_command))
+    subprocess.run(
+                untar_command,
+                capture_output=True,
+                check=False
+                )
